fet:add  team grid on the about page
